Revert "docs timer function migration"
